[repactor] store로 이벤트 이관
